+++ src/components/players/services.ts
@@ -3,10 +3,8 @@ import { AuthServices } from "../auth/services";
 import { PlayersDAO } from "@/db/players";
 import {
   Credentials,
-  PlayerOrderBy,
   PlayerRequest,
   PlayerUpdateRequest,
-  getPlayerId,
 } from "@/types/request/players";
 import { LoginResponse, PlainPlayerResponse } from "@/types/response/players";
 import { compare, hash } from "@/utils/crypt";
@@ -20,43 +18,12 @@ import { TokenDAO } from "@/db/token";
 import { Whatsapp } from "@/notification/whatsapp";
 import CONFIG, { PLAYER_STATUS } from "@/config";
 import { ForbiddenError } from "@/helpers/error";
+import { ResourceService } from "@/services/resource.service";
 
-export class PlayerServices {
-  /**
-   * @description Get all players.
-   * @returns Player[]
-   */
-  getAllPlayers = async (
-    page: number,
-    itemsPerPage: number,
-    search?: string,
-    orderBy?: PlayerOrderBy,
-  ): Promise<PlainPlayerResponse[]> => {
-    const players = await PlayersDAO._getAll(
-      page,
-      itemsPerPage,
-      search,
-      orderBy,
-    );
-
-    return players.map((player) => hidePassword(player));
-  };
-
-  getTotalPlayers = async (): Promise<number> => {
-    return await PlayersDAO.count;
-  };
-
-  /**
-   * @description Get player information by ID.
-   * @param playerId ID of the player to retrieve information.
-   * @returns Player
-   */
-  getPlayerById = async (playerId: getPlayerId): Promise<Player | null> => {
-    const player = await PlayersDAO._getById(playerId);
-    if (!player) return null;
-
-    return hidePassword<Player>(player);
-  };
+export class PlayerServices extends ResourceService {
+  constructor() {
+    super(PlayersDAO);
+  }
 
   /**
    * Create player