+++ src/routes/agent.router.ts
@@ -5,8 +5,6 @@ import { AgentController } from "@/components/agent/controller";
 import { validateCredentials } from "@/components/players/validators";
 import {
   validateBankAccountUpdate,
-  validateDepositIndex,
-  validateDepositUpdate,
   validateOnCallRequest,
   validatePaymentIndex,
   validateResetPasswordRequest,
@@ -14,7 +12,6 @@ import {
 } from "@/components/agent/validators";
 import { throwIfBadRequest } from "@/middlewares/requestErrorHandler";
 import { requireAgentRole } from "@/middlewares/auth";
-import { DepositController } from "@/components/deposits/controller";
 import { paymentRateLimiter } from "@/middlewares/rate-limiters/payment";
 
 const agentRouter = Router();
@@ -38,18 +35,6 @@ agentRouter.post(
   throwIfBadRequest,
   paymentRateLimiter,
   AgentController.releasePayment,
-);
-agentRouter.get(
-  "/deposits/:id?",
-  validateDepositIndex(),
-  throwIfBadRequest,
-  AgentController.showDeposits,
-);
-agentRouter.post(
-  "/deposits/:id",
-  validateDepositUpdate(),
-  throwIfBadRequest,
-  DepositController.create,
 );
 agentRouter.get("/bank-account", AgentController.getBankAccount);
 agentRouter.post(